+++ src/pages/index.tsx
@@ -1,243 +1,10 @@
-import { Layout } from "@/components/layout";
-import { Button } from "@/components/ui/button";
-
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-
-import { ToastAction } from "@/components/ui/toast";
-import { useToast } from "@/components/ui/use-toast";
-
-import { motion } from "framer-motion";
-import { Input } from "@/components/ui/input";
-import { cn } from "@/lib/utils";
-import { useForm } from "react-hook-form";
-import { registerSchema } from "@/validators/auth";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { z } from "zod";
-import { useState } from "react";
-import { ArrowRight } from "lucide-react";
-
-type RegisterType = z.infer<typeof registerSchema>;
+import { Layout } from "@/components/layout/Layout";
+import { SignInForm } from "@/pages/sign/SignInForm";
 
 export default function Home() {
-  const [step, setStep] = useState<number>(0);
-
-  const { toast } = useToast();
-
-  const form = useForm<RegisterType>({
-    resolver: zodResolver(registerSchema),
-    defaultValues: {
-      username: "",
-      email: "",
-      phone: "",
-      role: "",
-      password: "",
-      confirmPassword: "",
-    },
-  });
-
-  const onSubmit = (data: RegisterType) => {
-    const { password, confirmPassword } = data;
-    if (password !== confirmPassword) {
-      toast({
-        title: "비밀번호가 일치하지 않습니다.",
-        variant: "destructive",
-        description: "비밀번호를 다시 입력하시겠습니까?",
-        action: <ToastAction altText="Try again">Try again</ToastAction>,
-      });
-      return;
-    }
-    alert(JSON.stringify(data, null, 4));
-  };
-  const formValidationHandler = () => {
-    form.trigger(["username", "email", "phone", "role"]);
-    const usernameState = form.getFieldState("username");
-    const emailState = form.getFieldState("email");
-    const phoneState = form.getFieldState("phone");
-    const roleState = form.getFieldState("role");
-
-    if (!usernameState.isDirty || usernameState.invalid) return;
-    if (!emailState.isDirty || emailState.invalid) return;
-    if (!phoneState.isDirty || phoneState.invalid) return;
-    if (!roleState.isDirty || roleState.invalid) return;
-
-    setStep(1);
-  };
-  const pasteHandler = (event: React.ClipboardEvent<HTMLInputElement>) => {
-    event.preventDefault();
-    toast({
-      title: "붙여넣기를 할 수 없습니다.",
-      variant: "destructive",
-    });
-  };
   return (
     <Layout>
-      <Card className={cn("w-[400px]")}>
-        <CardHeader>
-          <CardTitle>회원가입</CardTitle>
-          <CardDescription>필수 정보를 입력해주세요.</CardDescription>
-        </CardHeader>
-        <CardContent>
-          <Form {...form}>
-            <form
-              onSubmit={form.handleSubmit(onSubmit)}
-              className="relative space-y-3 overflow-x-hidden"
-            >
-              <motion.div
-                className={cn("space-y-3")}
-                animate={{ translateX: `${step * -100}%` }}
-                transition={{ ease: "easeInOut" }}
-              >
-                <FormField
-                  control={form.control}
-                  name="username"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>이름</FormLabel>
-                      <FormControl>
-                        <Input placeholder="이름을 입력하세요" {...field} />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={form.control}
-                  name="email"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>이메일</FormLabel>
-                      <FormControl>
-                        <Input placeholder="이메일을 입력하세요" {...field} />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={form.control}
-                  name="phone"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>연락처</FormLabel>
-                      <FormControl>
-                        <Input placeholder="연락처를 입력하세요" {...field} />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={form.control}
-                  name="role"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>역할</FormLabel>
-                      <Select
-                        onValueChange={field.onChange}
-                        defaultValue={field.value}
-                      >
-                        <FormControl>
-                          <SelectTrigger>
-                            <SelectValue placeholder="역할을 선택해주세요" />
-                          </SelectTrigger>
-                        </FormControl>
-                        <SelectContent>
-                          <SelectItem value="admin">관리자</SelectItem>
-                          <SelectItem value="user">사용자</SelectItem>
-                        </SelectContent>
-                      </Select>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </motion.div>
-              <motion.div
-                className={cn("space-y-3 absolute top-0 left-0 right-0")}
-                animate={{ translateX: `${(1 - step) * 100}%` }}
-                style={{ translateX: `${(1 - step) * 100}%` }}
-                transition={{
-                  ease: "easeInOut",
-                }}
-              >
-                <FormField
-                  control={form.control}
-                  name="password"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>비밀번호</FormLabel>
-                      <FormControl>
-                        <Input type={"password"} {...field} />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={form.control}
-                  name="confirmPassword"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>비밀번호 확인</FormLabel>
-                      <FormControl>
-                        <Input
-                          type={"password"}
-                          {...field}
-                          onPaste={(event) => pasteHandler(event)}
-                        />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </motion.div>
-              <div className={"flex gap-2"}>
-                <Button className={cn({ hidden: step === 0 })} type="submit">
-                  계정 등록하기
-                </Button>
-                <Button
-                  type="submit"
-                  className={cn({ hidden: step === 1 })}
-                  onClick={formValidationHandler}
-                >
-                  다음 단계로
-                  <ArrowRight className="w-4 h-4 ml-2" />
-                </Button>
-                <Button
-                  type="button"
-                  variant={"ghost"}
-                  className={cn({ hidden: step === 0 })}
-                  onClick={() => {
-                    setStep(0);
-                  }}
-                >
-                  이전 단계로
-                </Button>
-              </div>
-            </form>
-          </Form>
-        </CardContent>
-      </Card>
+      <SignInForm />
     </Layout>
   );
 }
