+++ src/pages/sign/SignUpForm.tsx
@@ -1,0 +1,289 @@
+import { Button } from "@/components/ui/button";
+import { Layout } from "@/components/layout/Layout";
+
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
+
+import { ToastAction } from "@/components/ui/toast";
+import { useToast } from "@/components/ui/use-toast";
+
+import { motion } from "framer-motion";
+import { Input } from "@/components/ui/input";
+import { cn } from "@/lib/utils";
+import { useForm } from "react-hook-form";
+import { registerSchema } from "@/validators/auth";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { z } from "zod";
+import { useState } from "react";
+import { ArrowRight } from "lucide-react";
+import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group";
+import { useRouter } from "next/router";
+
+type RegisterType = z.infer<typeof registerSchema>;
+
+const SignUpForm = () => {
+  const [step, setStep] = useState<number>(0);
+  const { toast } = useToast();
+  const router = useRouter();
+
+  const form = useForm<RegisterType>({
+    resolver: zodResolver(registerSchema),
+    defaultValues: {
+      username: "",
+      email: "",
+      phone: "",
+      role: "",
+      gender: "",
+      password: "",
+      confirmPassword: "",
+    },
+  });
+  console.log("form :", form.watch());
+
+  const onSubmit = (data: RegisterType) => {
+    const { password, confirmPassword } = data;
+    if (password !== confirmPassword) {
+      toast({
+        title: "비밀번호가 일치하지 않습니다.",
+        variant: "destructive",
+        description: "비밀번호를 다시 입력하시겠습니까?",
+        action: <ToastAction altText="Try again">Try again</ToastAction>,
+      });
+      return;
+    }
+    alert(JSON.stringify(data, null, 4));
+  };
+  const formValidationHandler = () => {
+    form.trigger(["username", "email", "phone", "role", "gender"]);
+    const usernameState = form.getFieldState("username");
+    const emailState = form.getFieldState("email");
+    const phoneState = form.getFieldState("phone");
+    const roleState = form.getFieldState("role");
+    const genderState = form.getFieldState("gender");
+
+    if (!usernameState.isDirty || usernameState.invalid) return;
+    if (!emailState.isDirty || emailState.invalid) return;
+    if (!phoneState.isDirty || phoneState.invalid) return;
+    if (!roleState.isDirty || roleState.invalid) return;
+    if (!genderState.isDirty || genderState.invalid) return;
+
+    setStep(1);
+  };
+  const pasteHandler = (event: React.ClipboardEvent<HTMLInputElement>) => {
+    event.preventDefault();
+    toast({
+      title: "붙여넣기를 할 수 없습니다.",
+      variant: "destructive",
+    });
+  };
+  return (
+    <Layout>
+      <Card className={cn("w-[400px]")}>
+        <CardHeader>
+          <CardTitle>회원가입</CardTitle>
+          <CardDescription>필수 정보를 입력해주세요.</CardDescription>
+        </CardHeader>
+        <CardContent>
+          <Form {...form}>
+            <form
+              onSubmit={form.handleSubmit(onSubmit)}
+              className="relative space-y-3 overflow-x-hidden"
+            >
+              <motion.div
+                className={cn("space-y-3")}
+                animate={{ translateX: `${step * -100}%` }}
+                transition={{ ease: "easeInOut" }}
+              >
+                <FormField
+                  control={form.control}
+                  name="username"
+                  render={({ field }) => (
+                    <FormItem>
+                      <FormLabel>이름</FormLabel>
+                      <FormControl>
+                        <Input placeholder="이름을 입력하세요" {...field} />
+                      </FormControl>
+                      <FormMessage />
+                    </FormItem>
+                  )}
+                />
+                <FormField
+                  control={form.control}
+                  name="email"
+                  render={({ field }) => (
+                    <FormItem>
+                      <FormLabel>이메일</FormLabel>
+                      <FormControl>
+                        <Input placeholder="이메일을 입력하세요" {...field} />
+                      </FormControl>
+                      <FormMessage />
+                    </FormItem>
+                  )}
+                />
+                <FormField
+                  control={form.control}
+                  name="phone"
+                  render={({ field }) => (
+                    <FormItem>
+                      <FormLabel>연락처</FormLabel>
+                      <FormControl>
+                        <Input placeholder="010-XXXX-XXXX" {...field} />
+                      </FormControl>
+                      <FormMessage />
+                    </FormItem>
+                  )}
+                />
+                <FormField
+                  control={form.control}
+                  name="gender"
+                  render={({ field }) => (
+                    <FormItem className="space-y-3">
+                      <FormLabel>성별</FormLabel>
+                      <FormControl>
+                        <RadioGroup
+                          onValueChange={field.onChange}
+                          defaultValue={field.value}
+                          className="flex flex-col space-y-1"
+                        >
+                          <FormItem className="flex items-center space-x-1 space-y-0">
+                            <FormControl>
+                              <RadioGroupItem value="남성" />
+                            </FormControl>
+                            <FormLabel className="font-normal">남</FormLabel>
+                          </FormItem>
+                          <FormItem className="flex items-center space-x-1 space-y-0">
+                            <FormControl>
+                              <RadioGroupItem value="여성" />
+                            </FormControl>
+                            <FormLabel className="font-normal">녀</FormLabel>
+                          </FormItem>
+                        </RadioGroup>
+                      </FormControl>
+                      <FormMessage />
+                    </FormItem>
+                  )}
+                />
+                <FormField
+                  control={form.control}
+                  name="role"
+                  render={({ field }) => (
+                    <FormItem>
+                      <FormLabel>역할</FormLabel>
+                      <Select
+                        onValueChange={field.onChange}
+                        defaultValue={field.value}
+                      >
+                        <FormControl>
+                          <SelectTrigger>
+                            <SelectValue placeholder="역할을 선택해주세요" />
+                          </SelectTrigger>
+                        </FormControl>
+                        <SelectContent>
+                          <SelectItem value="admin">관리자</SelectItem>
+                          <SelectItem value="user">사용자</SelectItem>
+                        </SelectContent>
+                      </Select>
+                      <FormMessage />
+                    </FormItem>
+                  )}
+                />
+              </motion.div>
+              <motion.div
+                className={cn("space-y-3 absolute top-0 left-0 right-0")}
+                animate={{ translateX: `${(1 - step) * 100}%` }}
+                style={{ translateX: `${(1 - step) * 100}%` }}
+                transition={{
+                  ease: "easeInOut",
+                }}
+              >
+                <FormField
+                  control={form.control}
+                  name="password"
+                  render={({ field }) => (
+                    <FormItem>
+                      <FormLabel>비밀번호</FormLabel>
+                      <FormControl>
+                        <Input type={"password"} {...field} />
+                      </FormControl>
+                      <FormMessage />
+                    </FormItem>
+                  )}
+                />
+                <FormField
+                  control={form.control}
+                  name="confirmPassword"
+                  render={({ field }) => (
+                    <FormItem>
+                      <FormLabel>비밀번호 확인</FormLabel>
+                      <FormControl>
+                        <Input
+                          type={"password"}
+                          {...field}
+                          onPaste={(event) => pasteHandler(event)}
+                        />
+                      </FormControl>
+                      <FormMessage />
+                    </FormItem>
+                  )}
+                />
+              </motion.div>
+              <div className={"flex justify-between gap-2 pt-5"}>
+                <Button
+                  type="button"
+                  className={cn({ hidden: step === 1 })}
+                  onClick={formValidationHandler}
+                >
+                  다음 단계로
+                  <ArrowRight className="w-4 h-4 ml-2" />
+                </Button>
+                <Button
+                  className={cn({ hidden: step === 1 })}
+                  type="button"
+                  variant={"outline"}
+                  onClick={() => router.push("/")}
+                >
+                  로그인 페이지로
+                </Button>
+                <Button className={cn({ hidden: step === 0 })} type="submit">
+                  계정 등록하기
+                </Button>
+                <Button
+                  type="button"
+                  variant={"outline"}
+                  className={cn({ hidden: step === 0 })}
+                  onClick={() => {
+                    setStep(0);
+                  }}
+                >
+                  이전 단계로
+                </Button>
+              </div>
+            </form>
+          </Form>
+        </CardContent>
+      </Card>
+    </Layout>
+  );
+};
+
+export default SignUpForm;
